backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.0
langchain-google-genai==0.0.6
requests==2.31.0
python-dateutil==2.8.2
Pillow==10.0.1
django-filter==23.3
